[Mojo Blob URLs] Make clicking an <a target="blank_"> with a blob URL work correctly.
This goes through a code-path where the blob URL token was incorrectly
dropped, so make sure we copy the token from the old FrameLoadRequest to
the new FrameLoadRequest.

Adds a web-platform-test to verify this behavior, and enables the blob URL browser
tests to also run with MojoBlobURLs enabled. This browser test required another
bug fix in RenderFrameHostImpl.

Cq-Include-Trybots: luci.chromium.try:linux_mojo
Bug: 800901
Change-Id: I87eee70e784dd8cb273f8671f37b00d8a93d9543